% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs_nn.R
\name{labs_nn}
\alias{labs_nn}
\title{Use Nicknames in ggplot2 Plots}
\usage{
labs_nn(dict = "default")
}
\arguments{
\item{dict}{The dictionary name to use for nickname lookups (optional).}
}
\description{
This function provides a convenient way to apply human readable labels to
ggplot2 plots, by first registering them using \code{\link{nn_register}} and
then applying them using this function.
}
\details{
This makes it easy to specify nice names once and use them across a project.
}
\examples{
library(ggplot2)

# Register nicknames
nn_register(c(
  "mpg" = "Miles per Gallon",
  "hp" = "Horsepower",
  "factor(cyl)" = "Number of\nCylinders"
))

# Create plot and apply nickname labels
ggplot(mtcars, aes(x = mpg, y = hp, color = factor(cyl))) +
  geom_point() +
  labs_nn()

}
\seealso{
\code{\link{labs_map}} for direct remapping
}
